requirements.txt
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
pyodbc>=5.0.0
Flask>=3.0.0
Flask-Login>=0.6.3
Flask-SQLAlchemy>=3.1.0
Flask-Migrate>=4.0.0
python-dateutil>=2.8.0
openpyxl>=3.1.0
fpdf2>=2.7.0
gunicorn>=21.0.0
